requirements.txt
google-generativeai==0.5.2
python-dotenv==1.0.1
Pillow==10.3.0
pillow-heif==0.16.0
